feat: move the reporting limits to the global settings
Changelog: None
Ticket: None
